requirements.txt
streamlit
google-genai
python-dotenv
srt
pillow